Added purging of Atom and RDF feeds
This pull request is related to my question in the forum on rtCamp [1].

I’ve added lines for purging …/feed/atom/ and …/feed/rdf/ in addition to
…/feed/. The two RSS formats listed in the documentation [2] redirect to
…/feed/ on our system. Ideally, I imagine, one should somehow obtain a list
of feed URLs from Wordpress, since they may be configurable or may change
in the future.

As Pythonista, I’m not acquainted with PHP code style, but I tried to
imitate the style in the remainder of the file.

[1] http://rtcamp.com/support/topic/caching-feeds/
[2] http://codex.wordpress.org/WordPress_Feeds#Finding_Your_Feed_URL