Add feed purging to redis cache
For some reason it appears that the redis purger does not include purging feeds. This patch enables the same feed purging as exists for the fastCGI cache.

To make use of this purging, (especially if using easy engine) you may have to change your nginx config to allow caching feeds. In the stock configuration of the easy engine nginx configuration, feeds are not cached. On a popular site that can lead to a lot of unnecessary PHP processing.