Purge current page cache on front end
Change hook from 'admin_init' to 'admin_bar_init' so that purge_all() is called when purge button is clicked on both admin dashboard and front end admin bar.
Check if the user is on front end when set the call purge_url() and purge the current page cache on case basis.