Update class-nginx-helper-admin.php
Add an action that fires after the current URL has been purged.

There is already an action that fires after all URLs have been purged, but there isn't one for when the current page has been purged. This new action is useful when using a caching plugin such as Cloudflare to purge the current URL on Cloudflare as well.