Add support for the torden branch of `ngx_cache_purge`
The torden branch has support for `purge_all` with the following
directive added to the nginx configuration:

```
location = /purgeall {
	fastcgi_pass		unix:/dev/null;
	fastcgi_cache		WORDPRESS;
	fastcgi_cache_purge	PURGE purge_all from 127.0.0.1;
}
```

This makes it possible to purge the cache on systems where Nginx is
running as a different user to the PHP process.

Note, the `unix:/dev/null` in the above example is not a placeholder,
the module needs `fastcgi_pass` set to something valid or it will not
function, even though it never uses the value provided.